Bump three + @types/three 0.169 → 0.185 (#44)
Dependabot's grouped PR (#36) bumped three and @types/three together but failed
typecheck (and, on its stale pre-vite8/.npmrc base, smoke). The only real breakage
is a type tightening: @types/three 0.185 makes `DataTexture.image.data` nullable,
so the atlasTexture oracle's `tex.image.data.length` read no longer narrows.

- three / @types/three 0.169.0 → 0.185.0
- atlasTexture.test.ts: cast `tex.image.data as Uint8Array` before `.length`,
  matching the existing pattern already used a few lines up (line 19). No logic
  change — DataTexture still holds a plain Uint8Array.

Full gate verified locally on current main (vite 8 / vitest 4 / Stryker 9):
typecheck, 81 tests, lint, format:check, build, smoke (render census healthy),
and mutation:clean — score unchanged at 96.49%. Supersedes #36.

+++ src/render/atlasTexture.test.ts
@@ -57,7 +57,7 @@ describe("atlasTexture oracle", () => {
   test("the texture is the full atlas size and uses crisp (nearest) filtering", () => {
     const tex = buildAtlasTexture(8);
     expect(tex.image.width).toBe(ATLAS_COLS * 8);
-    expect(tex.image.data.length).toBe(tex.image.width * tex.image.height * 4);
+    expect((tex.image.data as Uint8Array).length).toBe(tex.image.width * tex.image.height * 4);
     expect(tex.magFilter).toBeDefined();
   });
 });
